requirements.txt
torch>=1.7.0,<2.0.0
torchvision>=0.8.1,<0.15.0
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
matplotlib>=3.2.0
seaborn>=0.11.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.7.0
